+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `itemsOrdination` mutation.
+- `updateItemsOrdination` mutation.
 
 ## [0.44.1] - 2020-10-06
 ### Fixed